+++ doc/changes.rst
@@ -8,6 +8,8 @@ New in 0.28
 
 * API changes:
 
+  - :class:`testnd.anova`: The ``match`` parameter is now determined
+    automatically and does not need to be specified anymore in most cases.
   - :attr:`testnd.ttest_1samp.diff` renamed to
     :attr:`testnd.ttest_1samp.difference`.
   - :class:`plot.Histogram`: following :mod:`matplotlib`, the ``normed``